+++ src/auth/controller/LoginController.ts
@@ -1,0 +1,21 @@
+/** @format */
+
+import JwtService from 'auth/service/JwtService';
+import LoginService from 'auth/service/LoginService';
+import {Request, Response} from 'express';
+
+const login =
+(loginService: LoginService, jwtService: JwtService) =>
+	async (req: Request, res: Response): Promise<void> => {
+		const {email, password} = req.body;
+		const {id, roles} = await loginService.login(email, password);
+
+		if (!id || !roles) {
+			res.status(401).json({error: 'Invalid credentials'});
+			return;
+		}
+
+		const token: string = jwtService.generateToken({userId: id, roles: roles});
+		res.json({token});
+	};
+export default login
